+++ domain/src/main/java/woowacourse/shopping/model/Price.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.model
 
 data class Price(val value: Int) {
     init {
